Echo the arguments used for config.py for later reference
It is useful to be able to look at the settings used to configure an
eCMD build. This can specifically be used to reproduce a build in a
fresh repository. In order to allow equivalent setups, make sure we
don't lose what arguments the build was configured with by printing
them out to makefile.config.