Output correct offset from server debug for I2C with 8-byte offsets
Commit 172df96 (Enable 8-byte offsets for I2C) stores the 8-byte
offset in the 'address' variable, rather than the 'offset' variable
used for 4-byte offsets. Update the debug message to do the same check
to output the correct variable.

The correct printf formatting string for a 64-bit value varies between
systems (%016lX or %016llX, for example), so pull in the PRIx64 macro from
inttypes.h.

Signed-off-by: Ryan King <rpking@us.ibm.com>