Return the git hash for `croquery serverversion`
We accomplish this by adding a makefile target to generate a .C file
if the description of HEAD changes. Because the added header file
remains unchanged if the commit changes, this doesn't require a mass
rebuild of the other files. Only git_version.o needs to be rebuilt.

Instead of printing out the entire hash directly, we print out the
most recent tag, an abbreviated hash, and a -dirty flag if
needed. `croquery serverversion` now returns something similar to:

...
Server Version
Date: May  6 2019 20:24:14
Git Version: ver-14-15-1-24-gedf2004
ControlInstruction       : 4
...

If building from a .zip download, the .git directory doesn't exist and
we have no way of knowing what commit we are on. Output a blank
version string in this case. Due to how the makefile was updated, the
git_version.C target is run every time in this situation, while it
only runs if the hash changes when cloning via git.

Signed-off-by: Ryan King <rpking@us.ibm.com>