Added ECMD_DLL_FILE_FIXED support (#272)
* Added ECMD_DEFAULT_DLL_FILE support

This allows an at build time definition of the ECMD_DLL_FILE to
load.  This removes the need for the ECMD_DLL_FILE to be set in
the environment.

This is designed for static install situations where the location
of the libraries doesn't change.  However, it still supports the
use setting ECMD_DLL_FILE if needed.

Signed-off-by: Jason Albert <albertj@us.ibm.com>

* Updated naming to ECMD_DLL_FILE_FIXED

Also added a comment showing how it could be used at compile time

Signed-off-by: Jason Albert <albertj@us.ibm.com>