Force python 3 and updated logic around use of linux_distribution
The script was compatible with both python2 and python3 so we used
generic python statement to invoke the script.  However some
places (yocto), have put the hammer down and don't even have a
python->python3 link.  This should still work on the old distros
ecmd supports, but they will have to have python3

platform.linux_distribution has been depricated for a while and is
finally removed from python3.8.  This code will attempt to use it
platform if at all available, otherise fallback to the distro
module and tell the user if the module wasn't found.

Signed-off-by: Jason Albert <albertj@us.ibm.com>